requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
openai-whisper>=20231117
pydub>=0.25.1
ffmpeg-python>=0.2.0
